review fixes: #42 round 9 — detect multiplexing, widen IPC/miss math
Copilot round 7:

1. Perf group was opened pinned with fallback-to-unpinned, but
   when pinning failed the read_format was still plain
   PERF_FORMAT_GROUP — giving the kernel license to multiplex the
   group without us being able to detect it. cycles/iter, IPC,
   and cache-miss% would be silently under-reported proportional
   to how much of the window the group was off-hardware.

   Fix: also request PERF_FORMAT_TOTAL_TIME_ENABLED |
   PERF_FORMAT_TOTAL_TIME_RUNNING. The read buffer now includes
   time_enabled and time_running; when time_running <
   time_enabled the kernel multiplexed. Rather than scale the
   counts (scaled numbers are a noisier proxy for the true rate),
   we flip last_read_ok_ to false so Bench::run suppresses the
   whole epoch. Consistent with how we handle short reads already.

2. IPC and cache-miss% computed `total_inst * 100` and
   `total_cmiss * 1000` in u64. For very long benches those
   products could overflow — theoretical at real scales but
   bulletproofing the math is trivial. Widened the multiplication
   to unsigned __int128 (same pattern err_pct already uses).

Test plan: test_perf_counters still passes; bench_bench_demo
still prints the same shape of output, now with multiplexing
detection active behind the scenes.

+++ testing/bench.h
@@ -370,8 +370,12 @@ struct Bench {
                 out_u64(total_inst / it);
             }
             if (pc.has(kPerfCycles) && pc.has(kPerfInstructions) && total_cycles > 0) {
-                // IPC × 100 (fixed-point 2 decimals)
-                const u64 ipc100 = (total_inst * 100) / total_cycles;
+                // IPC × 100 (fixed-point 2 decimals). Widen to
+                // __int128 for the multiply so long benches where
+                // total_inst exceeds ~1.8e17 don't overflow u64 and
+                // wrap to a nonsense IPC. Same pattern as err_pct.
+                const u64 ipc100 = static_cast<u64>(
+                    (static_cast<unsigned __int128>(total_inst) * 100) / total_cycles);
                 out("  IPC: ");
                 out_u64(ipc100 / 100);
                 out(".");
@@ -385,7 +389,8 @@ struct Bench {
                 // the kernel maps these to last-level-cache (LLC)
                 // counters, not L1 — label accordingly to avoid a
                 // misleading "L1-miss" attribution.
-                const u64 miss_thou = (total_cmiss * 1000) / total_cref;
+                const u64 miss_thou = static_cast<u64>(
+                    (static_cast<unsigned __int128>(total_cmiss) * 1000) / total_cref);
                 out("  cache-miss: ");
                 out_u64(miss_thou / 10);
                 out(".");
